vp: backup commit before start of vp web part
